% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{pokedex.sim}
\alias{pokedex.sim}
\title{Pokedex Simulator -Samples random pokemon}
\usage{
pokedex.sim(n, replace = T, gen = 1, prob)
}
\arguments{
\item{n}{number of pokemon to return}

\item{replace}{controls whether to sample with replacement or not}

\item{gen}{controls which generation of pokemon to sample from, default is 1st gen}

\item{prob}{an optional parameter that controls probability of sampling see sample()}
}
\value{
a dataframe of random selected pokemon
}
\description{
Pokedex Simulator -Samples random pokemon
}
\examples{
pokedex.sim(10)
}
