% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{pokemon.pop.sim}
\alias{pokemon.pop.sim}
\title{Pokemon Population Sampler - Samples specified pokemon and traits}
\usage{
pokemon.pop.sim(pokemon, n, traits, vary, sd = 1)
}
\arguments{
\item{pokemon}{name of pokemon in title case}

\item{n}{number of pokemon to return}

\item{traits}{column name from pokedex describing pokemon traits}

\item{vary}{controls whether to vary numerical trait values}

\item{sd}{optional parameter to influence how much variation to add}
}
\value{
a dataframe of pokemon with selected traits
}
\description{
Pokemon Population Sampler - Samples specified pokemon and traits
}
\examples{
pokemon.pop.sim("Bulbasaur", 10)
}
